modify/#125: 유저가 커넥션에 속하는지 확인하는 로직 수정

+++ src/features/user/user-connection/commands/disconnect-user-connection/disconnect-user-connection.command-handler.ts
@@ -45,6 +45,12 @@ export class DisconnectUserConnectionCommandHandler
       });
     }
 
+    if (!userConnection.isPartOfConnection(userId)) {
+      throw new HttpNotFoundException({
+        code: COMMON_ERROR_CODE.PERMISSION_DENIED,
+      });
+    }
+
     userConnection.disconnectConnection();
 
     await this.userConnectionRepository.delete(userConnection);
